requirements.txt
requests
logging
pillow